[PRDP-268] feat: refactored to define indipendent function class
