+++ src/main/java/it/gov/pagopa/receipt/pdf/helpdesk/utils/RecoverNotNotifiedReceiptUtils.java
@@ -1,7 +1,12 @@
 package it.gov.pagopa.receipt.pdf.helpdesk.utils;
 
+import com.azure.cosmos.models.FeedResponse;
 import it.gov.pagopa.receipt.pdf.helpdesk.entity.receipt.Receipt;
 import it.gov.pagopa.receipt.pdf.helpdesk.entity.receipt.enumeration.ReceiptStatusType;
+import it.gov.pagopa.receipt.pdf.helpdesk.service.ReceiptCosmosService;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class RecoverNotNotifiedReceiptUtils {
 
@@ -17,6 +22,24 @@ public class RecoverNotNotifiedReceiptUtils {
             receipt.setReasonErrPayer(null);
         }
         return receipt;
+    }
+
+    public static List<Receipt> receiptMassiveRestore(ReceiptStatusType statusType, ReceiptCosmosService receiptCosmosService) {
+        List<Receipt> receiptList = new ArrayList<>();
+        String continuationToken = null;
+        do {
+            Iterable<FeedResponse<Receipt>> feedResponseIterator =
+                    receiptCosmosService.getNotNotifiedReceiptByStatus(continuationToken, 100, statusType);
+
+            for (FeedResponse<Receipt> page : feedResponseIterator) {
+                for (Receipt receipt : page.getResults()) {
+                    Receipt restoredReceipt = restoreReceipt(receipt);
+                    receiptList.add(restoredReceipt);
+                }
+                continuationToken = page.getContinuationToken();
+            }
+        } while (continuationToken != null);
+        return receiptList;
     }
 
     private RecoverNotNotifiedReceiptUtils() {}
