+++ src/test/java/it/gov/pagopa/receipt/pdf/helpdesk/RecoverNotNotifiedReceiptMassiveTest.java
@@ -260,38 +260,7 @@ class RecoverNotNotifiedReceiptMassiveTest {
         verify(documentReceipts, never()).setValue(receiptCaptor.capture());
     }
 
-    @Test
-    public void scheduledTriggerShouldReturnAllValidReceiptsProcessed() {
-        FeedResponse feedResponseMock = mock(FeedResponse.class);
-        List<Receipt> receiptList = getReceiptList(ReceiptStatusType.IO_ERROR_TO_NOTIFY);
-        when(feedResponseMock.getResults()).thenReturn(receiptList);
-        when(receiptCosmosServiceMock.getNotNotifiedReceiptByStatus(any(), any(), eq(ReceiptStatusType.IO_ERROR_TO_NOTIFY)))
-                .thenReturn(Collections.singletonList(feedResponseMock));
 
-        FeedResponse feedResponseGenMock = mock(FeedResponse.class);
-        List<Receipt> receiptGenList = getReceiptList(ReceiptStatusType.GENERATED);
-        when(feedResponseGenMock.getResults()).thenReturn(receiptGenList);
-        when(receiptCosmosServiceMock.getNotNotifiedReceiptByStatus(any(), any(), eq(ReceiptStatusType.GENERATED)))
-                .thenReturn(Collections.singletonList(feedResponseGenMock));
-
-        sut.processRecoverNotNotifiedScheduledTrigger("info", documentReceipts, executionContextMock);
-
-        verify(documentReceipts).setValue(receiptCaptor.capture());
-
-        assertEquals(receiptList.size()+receiptGenList.size(), receiptCaptor.getValue().size());
-        Receipt captured1 = receiptCaptor.getValue().get(0);
-        assertEquals(ReceiptStatusType.GENERATED, captured1.getStatus());
-        assertEquals(EVENT_ID, captured1.getEventId());
-        assertEquals(0, captured1.getNotificationNumRetry());
-        assertNull(captured1.getReasonErr());
-        assertNull(captured1.getReasonErrPayer());
-        Receipt captured2 = receiptCaptor.getValue().get(1);
-        assertEquals(ReceiptStatusType.GENERATED, captured2.getStatus());
-        assertEquals(EVENT_ID, captured2.getEventId());
-        assertEquals(0, captured2.getNotificationNumRetry());
-        assertNull(captured2.getReasonErr());
-        assertNull(captured2.getReasonErrPayer());
-    }
 
     private Receipt buildReceipt(ReceiptStatusType statusType) {
         return Receipt.builder()
